requirements.txt
streamlit~=1.41.1
pillow~=11.1.0